requirements.txt
pycryptodome==3.21.0
pbkdf2==1.3